FDG-9789 clean up
