+++ src/components/data-preview/data-preview-section-container/data-preview-section-container.tsx
@@ -1,23 +1,16 @@
 import React, { FunctionComponent, useContext, useEffect, useMemo, useState } from 'react';
 import GLOBALS from '../../../helpers/constants';
-import {useSetRecoilState} from 'recoil';
-import {disableDownloadButtonState} from '../../../recoil/disableDownloadButtonState';
+import { useSetRecoilState } from 'recoil';
+import { disableDownloadButtonState } from '../../../recoil/disableDownloadButtonState';
 import moment from 'moment';
-import {
-  buildDateFilter,
-  buildSortParams,
-  fetchAllTableData,
-  fetchTableMeta,
-  formatDateForApi,
-  MAX_PAGE_SIZE
-} from '../../../utils/api-utils';
-import {queryClient} from '../../../../react-query-client';
-import {setTableConfig} from '../../dataset-data/table-section-container/set-table-config';
+import { buildDateFilter, buildSortParams, fetchAllTableData, fetchTableMeta, formatDateForApi, MAX_PAGE_SIZE } from '../../../utils/api-utils';
+import { queryClient } from '../../../../react-query-client';
+import { setTableConfig } from '../../dataset-data/table-section-container/set-table-config';
 import Analytics from '../../../utils/analytics/analytics';
-import {determineUserFilterUnmatchedForDateRange} from '../../filter-download-container/user-filter/user-filter';
-import {SetNoChartMessage} from '../../dataset-data/table-section-container/set-no-chart-message';
-import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {faSpinner} from '@fortawesome/free-solid-svg-icons';
+import { determineUserFilterUnmatchedForDateRange } from '../../filter-download-container/user-filter/user-filter';
+import { SetNoChartMessage } from '../../dataset-data/table-section-container/set-no-chart-message';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faSpinner } from '@fortawesome/free-solid-svg-icons';
 import AggregationNotice from '../../dataset-data/table-section-container/aggregation-notice/aggregation-notice';
 import SummaryTable from '../../dataset-data/table-section-container/summary-table/summary-table';
 import DataPreviewTable from '../data-preview-table/data-preview-table';
@@ -101,11 +94,9 @@ const DataPreviewSectionContainer: FunctionComponent<DataPreviewSectionProps> = 
   apiFilterDefault,
   setApiFilterDefault,
 }) => {
+  const { tableProps, setTableProps } = useContext(DataTableContext);
   const tableName = selectedTable.tableName;
   const [showPivotBar, setShowPivotBar] = useState(true);
-  const { tableProps, setTableProps, perPage, setPerPage, manualPagination, setManualPagination, reactTableSorting, setReactTableSort } = useContext(
-    DataTableContext
-  );
   const [legend, setLegend] = useState(window.innerWidth > GLOBALS.breakpoints.large);
   const [legendToggledByUser, setLegendToggledByUser] = useState(false);
   const [pivotsUpdated, setPivotsUpdated] = useState(false);
@@ -114,7 +105,10 @@ const DataPreviewSectionContainer: FunctionComponent<DataPreviewSectionProps> = 
   const [noChartMessage, setNoChartMessage] = useState(null);
   const [userFilterUnmatchedForDateRange, setUserFilterUnmatchedForDateRange] = useState(false);
   const [selectColumnPanel, setSelectColumnPanel] = useState(false);
+  const [perPage, setPerPage] = useState(null);
+  const [reactTableSorting, setReactTableSort] = useState([]);
   const [tableMeta, setTableMeta] = useState(null);
+  const [manualPagination, setManualPagination] = useState(false);
   const [apiErrorState, setApiError] = useState(apiError || false);
   const [chartData, setChartData] = useState(null);
 
@@ -403,7 +397,6 @@ const DataPreviewSectionContainer: FunctionComponent<DataPreviewSectionProps> = 
                     setSummaryValues={setSummaryValues}
                     pivotSelected={selectedPivot}
                     setSelectColumnPanel={setSelectColumnPanel}
-                    tableProps={tableProps}
                     selectedTable={selectedTable}
                     perPage={perPage}
                     setPerPage={setPerPage}
